Move unit tests into calculation sub-submodule
