+++ runtime/src/bank/partitioned_epoch_rewards/calculation.rs
@@ -156,7 +156,7 @@ impl Bank {
         }
     }
 
-    pub(in crate::bank) fn store_vote_accounts_partitioned(
+    fn store_vote_accounts_partitioned(
         &self,
         vote_account_rewards: VoteRewardsAccounts,
         metrics: &RewardsMetrics,
@@ -237,7 +237,7 @@ impl Bank {
     }
 
     /// Calculate epoch reward and return vote and stake rewards.
-    pub(in crate::bank) fn calculate_validator_rewards(
+    fn calculate_validator_rewards(
         &self,
         rewarded_epoch: Epoch,
         rewards: u64,
@@ -268,7 +268,7 @@ impl Bank {
 
     /// Calculates epoch reward points from stake/vote accounts.
     /// Returns reward lamports and points for the epoch or none if points == 0.
-    pub(in crate::bank) fn calculate_reward_points_partitioned(
+    fn calculate_reward_points_partitioned(
         &self,
         reward_calculate_params: &EpochRewardCalculateParamInfo,
         rewards: u64,
@@ -326,5 +326,314 @@ impl Bank {
         metrics.calculate_points_us.fetch_add(measure_us, Relaxed);
 
         (points > 0).then_some(PointValue { rewards, points })
+    }
+}
+
+#[cfg(test)]
+mod tests {
+    use {
+        super::*,
+        crate::{
+            bank::{null_tracer, tests::create_genesis_config, VoteReward},
+            genesis_utils::{
+                create_genesis_config_with_vote_accounts, GenesisConfigInfo, ValidatorVoteKeypairs,
+            },
+            stake_account::StakeAccount,
+            stakes::Stakes,
+        },
+        rayon::ThreadPoolBuilder,
+        solana_sdk::{
+            account::{accounts_equal, ReadableAccount, WritableAccount},
+            native_token::{sol_to_lamports, LAMPORTS_PER_SOL},
+            reward_type::RewardType,
+            signature::Signer,
+            stake::state::Delegation,
+            vote::state::{VoteStateVersions, MAX_LOCKOUT_HISTORY},
+        },
+        solana_vote_program::vote_state,
+        std::sync::RwLockReadGuard,
+    };
+
+    /// Helper function to create a bank that pays some rewards
+    fn create_reward_bank(expected_num_delegations: usize) -> (Bank, Vec<Pubkey>, Vec<Pubkey>) {
+        let validator_keypairs = (0..expected_num_delegations)
+            .map(|_| ValidatorVoteKeypairs::new_rand())
+            .collect::<Vec<_>>();
+
+        let GenesisConfigInfo { genesis_config, .. } = create_genesis_config_with_vote_accounts(
+            1_000_000_000,
+            &validator_keypairs,
+            vec![2_000_000_000; expected_num_delegations],
+        );
+
+        let bank = Bank::new_for_tests(&genesis_config);
+
+        // Fill bank_forks with banks with votes landing in the next slot
+        // Create enough banks such that vote account will root
+        for validator_vote_keypairs in &validator_keypairs {
+            let vote_id = validator_vote_keypairs.vote_keypair.pubkey();
+            let mut vote_account = bank.get_account(&vote_id).unwrap();
+            // generate some rewards
+            let mut vote_state = Some(vote_state::from(&vote_account).unwrap());
+            for i in 0..MAX_LOCKOUT_HISTORY + 42 {
+                if let Some(v) = vote_state.as_mut() {
+                    vote_state::process_slot_vote_unchecked(v, i as u64)
+                }
+                let versioned = VoteStateVersions::Current(Box::new(vote_state.take().unwrap()));
+                vote_state::to(&versioned, &mut vote_account).unwrap();
+                match versioned {
+                    VoteStateVersions::Current(v) => {
+                        vote_state = Some(*v);
+                    }
+                    _ => panic!("Has to be of type Current"),
+                };
+            }
+            bank.store_account_and_update_capitalization(&vote_id, &vote_account);
+        }
+        (
+            bank,
+            validator_keypairs
+                .iter()
+                .map(|k| k.vote_keypair.pubkey())
+                .collect(),
+            validator_keypairs
+                .iter()
+                .map(|k| k.stake_keypair.pubkey())
+                .collect(),
+        )
+    }
+
+    #[test]
+    fn test_store_vote_accounts_partitioned() {
+        let (genesis_config, _mint_keypair) = create_genesis_config(1_000_000 * LAMPORTS_PER_SOL);
+        let bank = Bank::new_for_tests(&genesis_config);
+
+        let expected_vote_rewards_num = 100;
+
+        let vote_rewards = (0..expected_vote_rewards_num)
+            .map(|_| Some((Pubkey::new_unique(), VoteReward::new_random())))
+            .collect::<Vec<_>>();
+
+        let mut vote_rewards_account = VoteRewardsAccounts::default();
+        vote_rewards.iter().for_each(|e| {
+            if let Some(p) = &e {
+                let info = RewardInfo {
+                    reward_type: RewardType::Voting,
+                    lamports: p.1.vote_rewards as i64,
+                    post_balance: p.1.vote_rewards,
+                    commission: Some(p.1.commission),
+                };
+                vote_rewards_account.rewards.push((p.0, info));
+                vote_rewards_account
+                    .accounts_to_store
+                    .push(e.as_ref().map(|p| p.1.vote_account.clone()));
+            }
+        });
+
+        let metrics = RewardsMetrics::default();
+
+        let stored_vote_accounts =
+            bank.store_vote_accounts_partitioned(vote_rewards_account, &metrics);
+        assert_eq!(expected_vote_rewards_num, stored_vote_accounts.len());
+
+        // load accounts to make sure they were stored correctly
+        vote_rewards.iter().for_each(|e| {
+            if let Some(p) = &e {
+                let (k, account) = (p.0, p.1.vote_account.clone());
+                let loaded_account = bank.load_slow_with_fixed_root(&bank.ancestors, &k).unwrap();
+                assert!(accounts_equal(&loaded_account.0, &account));
+            }
+        });
+    }
+
+    #[test]
+    fn test_store_vote_accounts_partitioned_empty() {
+        let (genesis_config, _mint_keypair) = create_genesis_config(1_000_000 * LAMPORTS_PER_SOL);
+        let bank = Bank::new_for_tests(&genesis_config);
+
+        let expected = 0;
+        let vote_rewards = VoteRewardsAccounts::default();
+        let metrics = RewardsMetrics::default();
+
+        let stored_vote_accounts = bank.store_vote_accounts_partitioned(vote_rewards, &metrics);
+        assert_eq!(expected, stored_vote_accounts.len());
+    }
+
+    #[test]
+    /// Test rewards computation and partitioned rewards distribution at the epoch boundary
+    fn test_rewards_computation() {
+        solana_logger::setup();
+
+        let expected_num_delegations = 100;
+        let bank = create_reward_bank(expected_num_delegations).0;
+
+        // Calculate rewards
+        let thread_pool = ThreadPoolBuilder::new().num_threads(1).build().unwrap();
+        let mut rewards_metrics = RewardsMetrics::default();
+        let expected_rewards = 100_000_000_000;
+
+        let calculated_rewards = bank.calculate_validator_rewards(
+            1,
+            expected_rewards,
+            null_tracer(),
+            &thread_pool,
+            &mut rewards_metrics,
+        );
+
+        let vote_rewards = &calculated_rewards.as_ref().unwrap().0;
+        let stake_rewards = &calculated_rewards.as_ref().unwrap().1;
+
+        let total_vote_rewards: u64 = vote_rewards
+            .rewards
+            .iter()
+            .map(|reward| reward.1.lamports)
+            .sum::<i64>() as u64;
+
+        // assert that total rewards matches the sum of vote rewards and stake rewards
+        assert_eq!(
+            stake_rewards.total_stake_rewards_lamports + total_vote_rewards,
+            expected_rewards
+        );
+
+        // assert that number of stake rewards matches
+        assert_eq!(stake_rewards.stake_rewards.len(), expected_num_delegations);
+    }
+
+    #[test]
+    fn test_rewards_point_calculation() {
+        solana_logger::setup();
+
+        let expected_num_delegations = 100;
+        let (bank, _, _) = create_reward_bank(expected_num_delegations);
+
+        let thread_pool = ThreadPoolBuilder::new().num_threads(1).build().unwrap();
+        let rewards_metrics = RewardsMetrics::default();
+        let expected_rewards = 100_000_000_000;
+
+        let stakes: RwLockReadGuard<Stakes<StakeAccount<Delegation>>> = bank.stakes_cache.stakes();
+        let reward_calculate_param = bank.get_epoch_reward_calculate_param_info(&stakes);
+
+        let point_value = bank.calculate_reward_points_partitioned(
+            &reward_calculate_param,
+            expected_rewards,
+            &thread_pool,
+            &rewards_metrics,
+        );
+
+        assert!(point_value.is_some());
+        assert_eq!(point_value.as_ref().unwrap().rewards, expected_rewards);
+        assert_eq!(point_value.as_ref().unwrap().points, 8400000000000);
+    }
+
+    #[test]
+    fn test_rewards_point_calculation_empty() {
+        solana_logger::setup();
+
+        // bank with no rewards to distribute
+        let (genesis_config, _mint_keypair) = create_genesis_config(sol_to_lamports(1.0));
+        let bank = Bank::new_for_tests(&genesis_config);
+
+        let thread_pool = ThreadPoolBuilder::new().num_threads(1).build().unwrap();
+        let rewards_metrics: RewardsMetrics = RewardsMetrics::default();
+        let expected_rewards = 100_000_000_000;
+        let stakes: RwLockReadGuard<Stakes<StakeAccount<Delegation>>> = bank.stakes_cache.stakes();
+        let reward_calculate_param = bank.get_epoch_reward_calculate_param_info(&stakes);
+
+        let point_value = bank.calculate_reward_points_partitioned(
+            &reward_calculate_param,
+            expected_rewards,
+            &thread_pool,
+            &rewards_metrics,
+        );
+
+        assert!(point_value.is_none());
+    }
+
+    #[test]
+    fn test_calculate_stake_vote_rewards() {
+        solana_logger::setup();
+
+        let expected_num_delegations = 1;
+        let (bank, voters, stakers) = create_reward_bank(expected_num_delegations);
+
+        let vote_pubkey = voters.first().unwrap();
+        let mut vote_account = bank
+            .load_slow_with_fixed_root(&bank.ancestors, vote_pubkey)
+            .unwrap()
+            .0;
+
+        let stake_pubkey = stakers.first().unwrap();
+        let stake_account = bank
+            .load_slow_with_fixed_root(&bank.ancestors, stake_pubkey)
+            .unwrap()
+            .0;
+
+        let thread_pool = ThreadPoolBuilder::new().num_threads(1).build().unwrap();
+        let mut rewards_metrics = RewardsMetrics::default();
+
+        let point_value = PointValue {
+            rewards: 100000, // lamports to split
+            points: 1000,    // over these points
+        };
+        let tracer = |_event: &RewardCalculationEvent| {};
+        let reward_calc_tracer = Some(tracer);
+        let rewarded_epoch = bank.epoch();
+        let stakes: RwLockReadGuard<Stakes<StakeAccount<Delegation>>> = bank.stakes_cache.stakes();
+        let reward_calculate_param = bank.get_epoch_reward_calculate_param_info(&stakes);
+        let (vote_rewards_accounts, stake_reward_calculation) = bank.calculate_stake_vote_rewards(
+            &reward_calculate_param,
+            rewarded_epoch,
+            point_value,
+            &thread_pool,
+            reward_calc_tracer,
+            &mut rewards_metrics,
+        );
+
+        assert_eq!(
+            vote_rewards_accounts.rewards.len(),
+            vote_rewards_accounts.accounts_to_store.len()
+        );
+        assert_eq!(vote_rewards_accounts.rewards.len(), 1);
+        let rewards = &vote_rewards_accounts.rewards[0];
+        let account = &vote_rewards_accounts.accounts_to_store[0];
+        let vote_rewards = 0;
+        let commision = 0;
+        _ = vote_account.checked_add_lamports(vote_rewards);
+        assert_eq!(
+            account.as_ref().unwrap().lamports(),
+            vote_account.lamports()
+        );
+        assert!(accounts_equal(account.as_ref().unwrap(), &vote_account));
+        assert_eq!(
+            rewards.1,
+            RewardInfo {
+                reward_type: RewardType::Voting,
+                lamports: vote_rewards as i64,
+                post_balance: vote_account.lamports(),
+                commission: Some(commision),
+            }
+        );
+        assert_eq!(&rewards.0, vote_pubkey);
+
+        assert_eq!(stake_reward_calculation.stake_rewards.len(), 1);
+        assert_eq!(
+            &stake_reward_calculation.stake_rewards[0].stake_pubkey,
+            stake_pubkey
+        );
+
+        let original_stake_lamport = stake_account.lamports();
+        let rewards = stake_reward_calculation.stake_rewards[0]
+            .stake_reward_info
+            .lamports;
+        let expected_reward_info = RewardInfo {
+            reward_type: RewardType::Staking,
+            lamports: rewards,
+            post_balance: original_stake_lamport + rewards as u64,
+            commission: Some(commision),
+        };
+        assert_eq!(
+            stake_reward_calculation.stake_rewards[0].stake_reward_info,
+            expected_reward_info,
+        );
     }
 }
